requirements.txt

# -- lily assistant
-e git+https://github.com/cosphere-org/lily-assistant.git#egg=lily-assistant

# -- http client
requests==2.21.0

# -- communication with websockets
websocket-client==0.46.0
